Got rid of decoder-aware metadata handling
Sorry for the churn on this, but it became obvious that encoding metadata in sentences is the wrong way to go, and leaves open the potential to cause other problems. This commit reverts the changes of commit aa10be5b63c6c2537d6b761a9d3315ee9a8bf25f, essentially: metadata can no longer be passed in to the decoder be encoding it between bars that start a sentence.

Metadata for changing operational parameters, adding rules to the custom grammar, and so on, are now only available to the HTTP server, when passed in on a URL string with the parameters "meta=...". The commands currently supported are

- changing a decoder weight
- getting a list of weights
- adding a custom grammar rule
- removing a custom grammar rule
- listing all the custom grammar rules