Fixed rule sorting for phrase-based decoding
The rules were sorted backwards due to an earlier fix that got rid of Joshua's
negating of phrase-table variables. The effect was that all values were computed
correctly, but the decoder retained the bottom config.num_translation_options rules
for each source, instead of the top. This doesn't affect Hiero decoding because
those rule lists are not truncated (note: they should be).

(This raises an idea: decoding could be made a lot faster if we sorted rules
using only stateless feature functions, and then trimmed the list to the top
config.num_translation_options items. This would add some loss of quality, likely
negligible, and would save language model costs on lots of items that are unlikely
to be used.)