First BnEnHiero test working (rest following soon)
Added a script to convert grammar files. It needs to be run from $JOSHUA/joshua-core because it is not smart about fixing path names, and needs to look at the grammar files to figure out what class type to use. It could be rewritten better to actually build a JSON output that gets written; as it is, it just builds text. But it works!