changed logic of glue grammars and custom grammar
Previously, custom grammars were always created. Now, they are only instantiated if a grammar with the owner "custom" is found in the config file. A suitable warning is thrown with a helpful error message if the user tries to add a rule when no custom grammar was loaded.

A glue grammar is created inline if no grammar owned by "glue" was found. Previously, one was created if there were no grammars period. This new method is more robust. It will be nice in version 7 to get rid of the separate PhraseTable, consolidating that and MemoryBasedBatchGrammar. Phrase-based decoding doesn't have to worry about there being glue rules present since the glue rules for phrase-based are hard-coded and part of the search algorithm, instead of just being treated like generic rules (like in CKY).